requirements.txt
attrs==19.3.0
autopep8==1.5.3
azure-cognitiveservices-vision-face==0.4.1
azure-common==1.1.25
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.1
decorator==4.4.2
Flask==1.1.2
idna==2.10
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
msrest==0.6.17
oauthlib==3.1.0
packaging==20.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
